requirements.txt
selenium
selenium_stealth
pyyaml
tzdata
python-dateutil